semana10/dia4/perudemy/backend/requirements.txt
django==3.2
djangorestframework
mysqlclient
cloudinary
django-cors-headers

